Provisioning2/requirements.txt
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
components==0.0.6
crypto==1.4.1
cryptography==2.8
idna==2.8
Naked==0.1.31
pycparser==2.19
pyOpenSSL==19.1.0
PyYAML==5.3
requests==2.22.0
shellescape==3.4.1
six==1.14.0
suds-py3==1.3.4.0
urllib3==1.25.7
